make binary format an optional extra
