+++ README.rst
@@ -23,6 +23,16 @@ Install
 
     $ pip install aioprometheus
 
+The standard installation of aioprometheus provides metrics in the text
+format. aioprometheus can encode metrics into the binary format too but it
+requires the optional ``aioprometheus-binary-format`` package to also
+be installed. If you want to make use of the binary metrics encoder you will
+need to install ``aioprometheus-binary-format`` manually.
+
+.. code-block:: console
+
+    $ pip install aioprometheus-binary-format
+
 
 Example
 -------
