+++ aioprometheus/__init__.py
@@ -2,9 +2,11 @@
 from .collectors import (
     Collector, Counter, Gauge, Summary, Histogram)
 from .decorators import count_exceptions, inprogress, timer
-from .formats import (BinaryFormatter, TextFormatter)
-from .pusher import Pusher
+from . import formats
+from . import pusher
+from . import negotiator
 from .registry import Registry, CollectorRegistry
 from .service import Service
 
-__version__ = "16.08.07"
+
+__version__ = "17.04.08"
