+++ aioprometheus/formats/__init__.py
@@ -1,3 +1,9 @@
 
+import importlib.util
+from .base import IFormatter
 from .text import TextFormatter
-from .binary import BinaryFormatter
+
+binary_format_spec = importlib.util.find_spec("aioprometheus_binary_format")
+binary_format_available = binary_format_spec is not None
+if binary_format_available:
+    from aioprometheus_binary_format import BinaryFormatter
