+++ requirements.dev.txt
@@ -3,5 +3,5 @@ autopep8>=1.2.4
 pycodestyle>=2.0.0
 sphinx>=1.4.4
 coverage>=4.1
-mypy-lang>=0.4.3
+mypy
 typed_ast>=0.5.5
